spectra/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
numpy
rasterio
GDAL==3.8.4
shapely
geopandas
scipy
torch
torchvision
segmentation-models-pytorch
scikit-learn
sqlalchemy
geoalchemy2
psycopg2-binary
alembic
apscheduler
resend
httpx
pytest
